Refatorando agent e visual de parcelas
