+++ requirements.txt
@@ -1,55 +1,2 @@
-openai==0.27.7
-google-genai==1.9.0
-altair==5.5.0
-attrs==25.3.0
-beautifulsoup4==4.13.4
-blinker==1.9.0
-bs4==0.0.1
-cachetools==5.5.2
-certifi==2025.8.3
-charset-normalizer==3.4.2
-click==8.2.1
-gitdb==4.0.12
-GitPython==3.1.45
-greenlet==3.2.3
-idna==3.10
-Jinja2==3.1.6
-joblib==1.2.0
-jsonschema==4.25.0
-jsonschema-specifications==2025.4.1
-lxml
-MarkupSafe==3.0.2
-narwhals==2.0.1
-numpy==2.3.2
-packaging==24.2
-pandas==2.3.1
-pillow==11.3.0
-plotly==5.18.0
-protobuf==3.20.3
-psycopg2-binary==2.9.10
-pyarrow==20.0.0
-pydeck==0.9.1
-python-dateutil==2.8.2
-python-dotenv==1.0.0
-pytz==2025.2
-referencing==0.36.2
-requests==2.32.4
-rpds-py==0.26.0
-scikit-learn==1.7.1
-scipy==1.15.3
-six==1.17.0
-smmap==5.0.2
-soupsieve==2.7
-SQLAlchemy==2.0.20
-streamlit==1.44.0
-tenacity==9.1.2
-threadpoolctl==3.6.0
-toml==0.10.2
-tornado==6.5.1
-typing_extensions==4.14.1
-tzdata==2025.2
-urllib3==2.5.0
-watchdog==6.0.0
-xlrd==2.0.2
-xlsxwriter==3.2.5
-dotenv
+# Automatically generated by https://github.com/damnever/pigar.
+
